Refactor game state and keypad into separate files
